Very work in progress outline
This outline was based upon an outline I created over the Christmas break of my sophmore year in GIMM (Games, Interactive Media, and Mobile).
It's simply based upon what I currently have, and what I intend to work on. As I continue to create more content and works, this will likely change. But for now, this is where I'm at and what I expect to publish on my portfolio. I did not spend time organizing this as is, since I do already have the outline organized within a document. This is simply for website placement and ideas.